requirements.txt
anthropic>=0.28.0
google-api-python-client>=2.100.0
google-auth>=2.20.0
pyngrok>=7.0.0
fastapi>=0.110.0
uvicorn>=0.27.0
elevenlabs>=2.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
pyaudio>=0.2.14       # modo voz (micrófono + altavoces)
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.26.0
pydantic>=2.7.0
rich>=13.7.0          # consola bonita para modo test
